Feat/tests

+++ src/tests/cradle-account-factory.t.sol
@@ -1,28 +1,25 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
-import { Test } from "forge-std/Test.sol";
-import { Vm } from "forge-std/Vm.sol";
-import { CradleAccountFactory } from "../core/CradleAccountFactory.sol";
-import { CradleAccount } from "../core/CradleAccount.sol";
-import { AccessController } from "../core/AccessController.sol";
-import { MockHTS } from "./utils/MockHTS.sol";
+import {Test} from "forge-std/Test.sol";
+import {Vm} from "forge-std/Vm.sol";
+import {CradleAccountFactory} from "../core/CradleAccountFactory.sol";
+import {CradleAccount} from "../core/CradleAccount.sol";
+import {AccessController} from "../core/AccessController.sol";
+import {MockHTS} from "./utils/MockHTS.sol";
 
 contract CradleAccountFactoryTest is Test {
     CradleAccountFactory factory;
     AccessController acl;
     MockHTS mockHTS;
-    
+
     address admin;
     address user1;
     address user2;
     address constant HTS_PRECOMPILE = address(0x167);
 
     event AccountCreated(
-        address indexed accountAddress,
-        string indexed controller,
-        address indexed creator,
-        uint64 allowList
+        address indexed accountAddress, string indexed controller, address indexed creator, uint64 allowList
     );
     event AccountLinkedToUser(address indexed accountAddress, address indexed user);
     event AccountUnlinkedFromUser(address indexed accountAddress, address indexed user);
@@ -31,13 +28,13 @@ contract CradleAccountFactoryTest is Test {
         admin = address(this);
         user1 = makeAddr("user1");
         user2 = makeAddr("user2");
-        
+
         mockHTS = new MockHTS();
         vm.etch(HTS_PRECOMPILE, address(mockHTS).code);
-        
+
         acl = new AccessController();
         acl.grantAccess(1, admin);
-        
+
         factory = new CradleAccountFactory(address(acl), 1);
     }
 
@@ -53,7 +50,7 @@ contract CradleAccountFactoryTest is Test {
     // createAccount Tests
     function test_CreateAccount_CreatesAccountSuccessfully() public {
         address accountAddress = factory.createAccount("user123", 1);
-        
+
         assertNotEq(accountAddress, address(0));
         assertTrue(factory.isValidAccount(accountAddress));
     }
@@ -61,46 +58,41 @@ contract CradleAccountFactoryTest is Test {
     function test_CreateAccount_EmitsEvent() public {
         // We can't predict the exact address, so we check for event emission without strict address matching
         vm.recordLogs();
-        
+
         address accountAddress = factory.createAccount("user456", 1);
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
         bool foundEvent = false;
-        
-        for (uint i = 0; i < logs.length; i++) {
+
+        for (uint256 i = 0; i < logs.length; i++) {
             if (logs[i].topics[0] == keccak256("AccountCreated(address,string,address,uint64)")) {
                 foundEvent = true;
                 break;
             }
         }
-        
+
         assertTrue(foundEvent, "AccountCreated event not emitted");
         assertNotEq(accountAddress, address(0));
     }
 
     function test_CreateAccount_RegistersByController() public {
         address accountAddress = factory.createAccount("controller789", 1);
-        
+
         assertEq(factory.accountsByController("controller789"), accountAddress);
     }
 
     function test_CreateAccount_AddsToAllAccounts() public {
         address account1 = factory.createAccount("user1", 1);
         address account2 = factory.createAccount("user2", 1);
-        
+
         assertEq(factory.allAccounts(0), account1);
         assertEq(factory.allAccounts(1), account2);
     }
 
     function test_CreateAccount_RevertsIfControllerExists() public {
         factory.createAccount("duplicate", 1);
-        
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                CradleAccountFactory.AccountAlreadyExists.selector,
-                "duplicate"
-            )
-        );
+
+        vm.expectRevert(abi.encodeWithSelector(CradleAccountFactory.AccountAlreadyExists.selector, "duplicate"));
         factory.createAccount("duplicate", 1);
     }
 
@@ -113,7 +105,7 @@ contract CradleAccountFactoryTest is Test {
     // createAccountForUser Tests
     function test_CreateAccountForUser_CreatesAndLinksAccount() public {
         address accountAddress = factory.createAccountForUser("user_ctrl", user1, 1);
-        
+
         assertNotEq(accountAddress, address(0));
         assertEq(factory.accountsByUser(user1), accountAddress);
         assertEq(factory.accountsByController("user_ctrl"), accountAddress);
@@ -122,19 +114,14 @@ contract CradleAccountFactoryTest is Test {
     function test_CreateAccountForUser_EmitsBothEvents() public {
         vm.expectEmit(false, true, true, true);
         emit AccountCreated(address(0), "ctrl123", admin, 1);
-        
+
         factory.createAccountForUser("ctrl123", user1, 1);
     }
 
     function test_CreateAccountForUser_RevertsIfUserHasAccount() public {
         factory.createAccountForUser("first", user1, 1);
-        
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                CradleAccountFactory.UserAlreadyHasAccount.selector,
-                user1
-            )
-        );
+
+        vm.expectRevert(abi.encodeWithSelector(CradleAccountFactory.UserAlreadyHasAccount.selector, user1));
         factory.createAccountForUser("second", user1, 1);
     }
 
@@ -147,43 +134,33 @@ contract CradleAccountFactoryTest is Test {
     // linkAccountToUser Tests
     function test_LinkAccountToUser_LinksExistingAccount() public {
         address accountAddress = factory.createAccount("existing", 1);
-        
+
         vm.expectEmit(true, true, false, true);
         emit AccountLinkedToUser(accountAddress, user1);
-        
+
         factory.linkAccountToUser("existing", user1);
-        
+
         assertEq(factory.accountsByUser(user1), accountAddress);
     }
 
     function test_LinkAccountToUser_RevertsIfUserHasAccount() public {
         factory.createAccount("acc1", 1);
         factory.linkAccountToUser("acc1", user1);
-        
+
         factory.createAccount("acc2", 1);
-        
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                CradleAccountFactory.UserAlreadyHasAccount.selector,
-                user1
-            )
-        );
+
+        vm.expectRevert(abi.encodeWithSelector(CradleAccountFactory.UserAlreadyHasAccount.selector, user1));
         factory.linkAccountToUser("acc2", user1);
     }
 
     function test_LinkAccountToUser_RevertsIfAccountNotFound() public {
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                CradleAccountFactory.AccountNotFound.selector,
-                "nonexistent"
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(CradleAccountFactory.AccountNotFound.selector, "nonexistent"));
         factory.linkAccountToUser("nonexistent", user1);
     }
 
     function test_LinkAccountToUser_RevertsWithoutAuthorization() public {
         factory.createAccount("test", 1);
-        
+
         vm.prank(user1);
         vm.expectRevert("Unauthorized");
         factory.linkAccountToUser("test", user2);
@@ -192,37 +169,32 @@ contract CradleAccountFactoryTest is Test {
     // unlinkUserFromAccount Tests
     function test_UnlinkUserFromAccount_UnlinksAccount() public {
         address accountAddress = factory.createAccountForUser("user_test", user1, 1);
-        
+
         vm.expectEmit(true, true, false, true);
         emit AccountUnlinkedFromUser(accountAddress, user1);
-        
+
         factory.unlinkUserFromAccount(user1);
-        
+
         assertEq(factory.accountsByUser(user1), address(0));
     }
 
     function test_UnlinkUserFromAccount_RevertsWithoutAuthorization() public {
         factory.createAccountForUser("user_test", user1, 1);
-        
+
         vm.prank(user1);
         vm.expectRevert("Unauthorized");
         factory.unlinkUserFromAccount(user1);
     }
 
     function test_UnlinkUserFromAccount_RevertsIfUserHasNoAccount() public {
-        vm.expectRevert(
-            abi.encodeWithSelector(
-                CradleAccountFactory.InvalidAccount.selector,
-                address(0)
-            )
-        );
+        vm.expectRevert(abi.encodeWithSelector(CradleAccountFactory.InvalidAccount.selector, address(0)));
         factory.unlinkUserFromAccount(user1);
     }
 
     // getAccountByController Tests
     function test_GetAccountByController_ReturnsCorrectAccount() public {
         address accountAddress = factory.createAccount("finder", 1);
-        
+
         address found = factory.getAccountByController("finder");
         assertEq(found, accountAddress);
     }
@@ -243,10 +215,10 @@ contract CradleAccountFactoryTest is Test {
     function test_Integration_CreateMultipleAccountsAndLink() public {
         address account1 = factory.createAccount("ctrl1", 1);
         address account2 = factory.createAccount("ctrl2", 2);
-        
+
         factory.linkAccountToUser("ctrl1", user1);
         factory.linkAccountToUser("ctrl2", user2);
-        
+
         assertEq(factory.accountsByUser(user1), account1);
         assertEq(factory.accountsByUser(user2), account2);
         assertEq(factory.accountsByController("ctrl1"), account1);
@@ -255,10 +227,10 @@ contract CradleAccountFactoryTest is Test {
 
     function test_Integration_UnlinkAndRelink() public {
         address account1 = factory.createAccountForUser("user1_ctrl", user1, 1);
-        
+
         factory.unlinkUserFromAccount(user1);
         assertEq(factory.accountsByUser(user1), address(0));
-        
+
         // Can link another account now
         address account2 = factory.createAccount("user1_ctrl2", 1);
         factory.linkAccountToUser("user1_ctrl2", user1);
